move helper functions into a new file
